+++ src/main.rs
@@ -730,6 +730,18 @@ fn dispatch_common_subcommand(
             let result = commands::list_shovels(client);
             res_handler.tabular_result(result)
         }
+        ("streams", "declare") => {
+            let result = commands::declare_stream(client, &vhost, second_level_args);
+            res_handler.no_output_on_success(result);
+        }
+        ("streams", "delete") => {
+            let result = commands::delete_queue(client, &vhost, second_level_args);
+            res_handler.delete_operation_result(result);
+        }
+        ("streams", "list") => {
+            let result = commands::list_queues(client, &vhost);
+            res_handler.tabular_result(result)
+        }
         ("users", "connections") => {
             let result = commands::list_user_connections(client, second_level_args);
             res_handler.tabular_result(result)